Change the structure to better support testing
