+++ create_and_mint_case_using_stores.py
@@ -21,13 +21,6 @@ from create_mint_store import make_mint_store
 from flask import Flask
 from flask_restful import Api
 
-app = Flask(__name__)
-app.config['SQLALCHEMY_DATABASE_URI'] = 'postgres://sg:sg@localhost:8082/sg'
-
-init_database(app)
-
-api = Api(app)
-
 
 def apply_mintstore_to_case_field(mintstore, case_field):
     """
@@ -115,7 +108,7 @@ def recursively_get_case_fields_with_specs(case_field):
     return output_list
 
 
-def mint_case(name, case, user, mintstoremap={}):
+def mint_case(session, name, case, user, mintstoremap={}):
     """
     Turn a case into a job by minting it
     """
@@ -138,79 +131,92 @@ def mint_case(name, case, user, mintstoremap={}):
     return new_minted_case
 
 
-if __name__ == '__main__':
-    session = db.session
-
+def set_up_test_database():
+    """
+    Set up a database with some stuff in it
+    """
     # preamble - pretend this bit is being done before hand #################
 
     # add some cases to the Tank Store and the Fluid Store
+    session = db.session
     tank = make_tank_store()
     fluid = make_fluid_store()
 
+    done = False
+    for case in Case.query.all():
+        done = True
+    if done:
+        print("Data already there!")
+        exit()
+    session.add(tank)
+    session.add(fluid)
+    session.commit()
+    # add milk and tankX (fixed values for MintedCases) to the MintStore
+    make_mint_store(session)
+
+    # end of preamble #################
+
+    # create a real case, using the tank from the tank store
+    #  and the fluid from the fluid store
+
+    mycase = Case(name='MyCase')
+
+    new_tank_store = session.query(Case). \
+        filter(Case.name == 'tanks_R_us').first()
+
+    new_tankA = new_tank_store.fields[0].deep_copy()
+    new_fluid_store = session.query(Case). \
+        filter(Case.name == 'fluids_R_us').first()
+
+    new_fluid1 = new_fluid_store.fields[0].deep_copy()
+    new_fluid1.name = 'fluid 1'
+    new_fluid1.prepend_prefix('Fluid1_')
+    new_fluid2 = new_fluid_store.fields[0].deep_copy()
+    new_fluid2.name = 'fluid 2'
+    new_fluid2.prepend_prefix('Fluid2_')
+
+    mycase.fields.append(new_tankA)
+    mycase.fields.append(new_fluid1)
+    mycase.fields.append(new_fluid2)
+    # save this to the DB
+    session.add(mycase)
+    session.commit()
+
+    # test the prefix has been added
+    # print('prefix to fluid2 viscosity ',
+    #       mycase.fields[2].child_field[1].specs[-1].property_value)
+
+    ###################################################################
+    # ########################try minting a case ########################
+
+    # create a dictionary with the mapping between
+    # the case names and the mintstore names
+    # => i guess the UI will have a better way of doing this in real life!
+
+    mintstoremap = {
+        'tankA': 'TankX',
+        'fluid 1': 'Milk',
+        'fluid 2': 'Water'
+    }
+
+    minted_case = mint_case(session, 'TESTMINT',
+                            mycase, 'nbarlow', mintstoremap)
+    print('\n ==== Created a MintedCase ====== \n')
+
+    for v in minted_case.values:
+        print('Minted case has %s = %s' % (v.name, v.value))
+
+    session.add(minted_case)
+    session.commit()
+
+
+if __name__ == '__main__':
+    app = Flask(__name__)
+    app.config['SQLALCHEMY_DATABASE_URI'] = ('postgres://'
+                                             'sg:sg@localhost:8082/sg')
+
+    init_database(app)
+
+    api = Api(app)
     with app.app_context():
-
-        done = False
-        for case in Case.query.all():
-            done = True
-        if done:
-            print("Data already there!")
-            exit()
-        session.add(tank)
-        session.add(fluid)
-        session.commit()
-        # add milk and tankX (fixed values for MintedCases) to the MintStore
-        make_mint_store(session)
-
-        # end of preamble #################
-
-        # create a real case, using the tank from the tank store
-        #  and the fluid from the fluid store
-
-        mycase = Case(name='MyCase')
-
-        new_tank_store = session.query(Case). \
-            filter(Case.name == 'tanks_R_us').first()
-
-        new_tankA = new_tank_store.fields[0].deep_copy()
-        new_fluid_store = session.query(Case). \
-            filter(Case.name == 'fluids_R_us').first()
-
-        new_fluid1 = new_fluid_store.fields[0].deep_copy()
-        new_fluid1.name = 'fluid 1'
-        new_fluid1.prepend_prefix('Fluid1_')
-        new_fluid2 = new_fluid_store.fields[0].deep_copy()
-        new_fluid2.name = 'fluid 2'
-        new_fluid2.prepend_prefix('Fluid2_')
-
-        mycase.fields.append(new_tankA)
-        mycase.fields.append(new_fluid1)
-        mycase.fields.append(new_fluid2)
-        # save this to the DB
-        session.add(mycase)
-        session.commit()
-
-        # test the prefix has been added
-        # print('prefix to fluid2 viscosity ',
-        #       mycase.fields[2].child_field[1].specs[-1].property_value)
-
-        ###################################################################
-        # ########################try minting a case ########################
-
-        # create a dictionary with the mapping between
-        # the case names and the mintstore names
-        # => i guess the UI will have a better way of doing this in real life!
-
-        mintstoremap = {
-            'tankA': 'TankX',
-            'fluid 1': 'Milk',
-            'fluid 2': 'Water'
-        }
-
-        minted_case = mint_case('TESTMINT', mycase, 'nbarlow', mintstoremap)
-        print('\n ==== Created a MintedCase ====== \n')
-
-        for v in minted_case.values:
-            print('Minted case has %s = %s' % (v.name, v.value))
-
-        session.add(minted_case)
-        session.commit()
+        set_up_test_database()
